[Feat]#79 스캔뷰 리테이크 구현
